services/streamlit-ui/requirements.txt
streamlit==1.37.0
requests==2.32.3
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-requests